move sqls crate to pg crate submodule
