--- crates/sqls/src/lib.rs
@@ -1,10 +1,0 @@
-pub mod account;
-pub mod approval;
-pub mod balance;
-pub mod common;
-pub mod cte;
-pub mod profile;
-pub mod rule_instance;
-pub mod transaction;
-pub mod transaction_item;
-// cadet todo: increase test coverage. measure with "cargo llvm-cov"
